Update: SJF_NP works correctly according to priority
An old problem with SJF_NP is that it does not follow the priority. This is made that when the priority is a background process, it should be preempted still. (not sure if it should be preempted as it ruins the logical flow of a non-preemptive algorithm. This update is subject to change)